+++ tests/AdagucTests/TestGenericImageWarper.py
@@ -1,6 +1,11 @@
 import os
 import pytest
 from adaguc.AdagucTestTools import AdagucTestTools
+from conftest import (
+    make_adaguc_env,
+    run_adaguc_and_compare_image,
+    update_db,
+)
 
 ADAGUC_PATH = os.environ["ADAGUC_PATH"]
 
@@ -103,4 +108,28 @@ class TestGenericImageWarper:
             self.testresultspath + filetocheck,
             maxAllowedColorDifference=1,
             maxAllowedColorPercentage=0.02,
+        )
+
+    def test_getmap_latlon_latdecreasing_and_increasing_genericbilinear(self):
+        """
+        Test rendering of data aligned upwards and downwards with bilinear interpolation
+        """
+        env = make_adaguc_env("adaguc.tests.genericrender.xml", self.testresultspath, self.expectedoutputsspath)
+        update_db(env)
+        run_adaguc_and_compare_image(
+            env,
+            "test_getmap_latlon_latdecreasing_and_increasing_genericbilinear.png",
+            "DATASET=adaguc.tests.genericrender&SERVICE=WMS&&SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=latlon_latdecreasing_and_increasing_genericbilinear&WIDTH=256&HEIGHT=512&CRS=EPSG%3A3857&BBOX=471070.27730388555,6649427.860630986,982084.0646480684,7697681.648355679&STYLES=autobilinear&FORMAT=image/png&TRANSPARENT=TRUE&&time=2017-10-15T07%3A31%3A53Z&0.9458434770354436",
+        )
+
+    def test_getmap_latlon_latdecreasing_and_increasing_genericnearest(self):
+        """
+        Test rendering of data aligned upwards and downwards with nearest interpolation
+        """
+        env = make_adaguc_env("adaguc.tests.genericrender.xml", self.testresultspath, self.expectedoutputsspath)
+        update_db(env)
+        run_adaguc_and_compare_image(
+            env,
+            "test_getmap_latlon_latdecreasing_and_increasing_genericnearest.png",
+            "DATASET=adaguc.tests.genericrender&SERVICE=WMS&&SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=latlon_latdecreasing_and_increasing_genericnearest&WIDTH=256&HEIGHT=512&CRS=EPSG%3A3857&BBOX=471070.27730388555,6649427.860630986,982084.0646480684,7697681.648355679&STYLES=autogeneric&FORMAT=image/png&TRANSPARENT=TRUE&&time=2017-10-15T07%3A31%3A53Z&0.9458434770354436",
         )