+++ NEWS.md
@@ -1,3 +1,7 @@
+**Version 7.2.0 - TBD***
+
+- Fixed [#614](https://github.com/KNMI/adaguc-server/issues/614) where generic renderer caused issues with linear transform
+
 **Version 7.1.2 - 2026-03-30***
 
 - Refactored dataobject and datasource properties - no heap data is owned anymore.